# DSell_API_OpenAPI.json
{
    "openapi":  "3.0.3",
    "info":  {
                 "title":  "DSell 直销管理系统 API",
                 "version":  "1.0.0",
                 "description":  "Base URL: http://localhost:8080\n认证方式: Header 中携带 token (登录接口返回)"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080"
                    }
                ],
    "security":  [
                     {
                         "token":  [

                                   ]
                     }
                 ],
    "components":  {
                       "securitySchemes":  {
                                               "token":  {
                                                             "type":  "apiKey",
                                                             "in":  "header",
                                                             "name":  "token"
                                                         }
                                           },
                       "schemas":  {
...
                                       "WithdrawApplyDTO":  {
                                                                "type":  "object",
                                                                "required":  [
                                                                                 "bankName",
                                                                                 "bankCard"
                                                                             ],
                                                                "properties":  {
                                                                                   "amount":  {
                                                                                                  "type":  "number",
                                                                                                  "example":  500.00
                                                                                              },
                                                                                   "bankName":  {
                                                                                                    "type":  "string",
                                                                                                    "example":  "中国工商银行"
                                                                                                },
                                                                                   "bankCard":  {
                                                                                                    "type":  "string",
                                                                                                    "example":  "6222021234567890123"
                                                                                                }
                                                                               }
                                                            },
...
                                   }
                   },
...
    "paths":  {
...
                  "/api/finance/wallet/logs":  {
                                                   "get":  {
                                                               "tags":  [
                                                                            "Finance"
                                                                        ],
                                                               "summary":  "钱包流水 (管理端)",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "page",
                                                                                      "in":  "query",
                                                                                      "schema":  {
                                                                                                     "type":  "integer",
                                                                                                     "default":  1
                                                                                                 }
                                                                                  },
                                                                                  {
                                                                                      "name":  "size",
                                                                                      "in":  "query",
                                                                                      "schema":  {
                                                                                                     "type":  "integer",
                                                                                                     "default":  10
                                                                                                 }
                                                                                  }
                                                                              ]
                                                           }
                                               },
                  "/api/finance/wallet/logs/my":  {
                                                      "get":  {
                                                                  "tags":  [
                                                                               "Finance"
                                                                           ],
                                                                  "summary":  "我的钱包流水",
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "page",
                                                                                         "in":  "query",
                                                                                         "schema":  {
                                                                                                        "type":  "integer",
                                                                                                        "default":  1
                                                                                                    }
                                                                                     },
                                                                                     {
                                                                                         "name":  "size",
                                                                                         "in":  "query",
                                                                                         "schema":  {
                                                                                                        "type":  "integer",
                                                                                                        "default":  10
                                                                                                    }
                                                                                     }
                                                                                 ]
                                                              }
                                                  },
                  "/api/finance/withdraw/apply":  {
                                                      "post":  {
                                                                   "tags":  [
                                                                                "Finance"
                                                                            ],
                                                                   "summary":  "申请提现",
                                                                   "requestBody":  {
                                                                                       "required":  true,
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/WithdrawApplyDTO"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   }
                                                               }
                                                  },
                  "/api/finance/withdraw/page":  {
                                                     "get":  {
                                                                 "tags":  [
                                                                              "Finance"
                                                                          ],
                                                                 "summary":  "分页查询提现 (管理端)",
                                                                 "parameters":  [
                                                                                    {
                                                                                        "name":  "page",
                                                                                        "in":  "query",
                                                                                        "schema":  {
                                                                                                       "type":  "integer",
                                                                                                       "default":  1
                                                                                                   }
                                                                                    },
                                                                                    {
                                                                                        "name":  "size",
                                                                                        "in":  "query",
                                                                                        "schema":  {
                                                                                                       "type":  "integer",
                                                                                                       "default":  10
                                                                                                   }
                                                                                    },
                                                                                    {
                                                                                        "name":  "status",
                                                                                        "in":  "query",
                                                                                        "schema":  {
                                                                                                       "type":  "integer",
                                                                                                       "enum":  [
                                                                                                                    0,
                                                                                                                    1,
                                                                                                                    2
                                                                                                                ]
                                                                                                   },
                                                                                        "description":  "0=待审核 1=通过 2=拒绝"
                                                                                    }
                                                                                ]
                                                             }
                                                 },
                  "/api/finance/withdraw/my":  {
                                                   "get":  {
                                                               "tags":  [
                                                                            "Finance"
                                                                        ],
                                                               "summary":  "我的提现记录",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "page",
                                                                                      "in":  "query",
                                                                                      "schema":  {
                                                                                                     "type":  "integer",
                                                                                                     "default":  1
                                                                                                 }
                                                                                  },
                                                                                  {
                                                                                      "name":  "size",
                                                                                      "in":  "query",
                                                                                      "schema":  {
                                                                                                     "type":  "integer",
                                                                                                     "default":  10
                                                                                                 }
                                                                                  }
                                                                              ]
                                                           }
                                               },
                  "/api/finance/withdraw/{id}/audit":  {
                                                           "put":  {
                                                                       "tags":  [
                                                                                    "Finance"
                                                                                ],
                                                                       "summary":  "审核提现",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "id",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "integer"
                                                                                                         }
                                                                                          },
                                                                                          {
                                                                                              "name":  "status",
                                                                                              "in":  "query",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "integer",
                                                                                                             "enum":  [
                                                                                                                          1,
                                                                                                                          2
                                                                                                                      ]
                                                                                                         },
                                                                                              "description":  "1=通过 2=拒绝"
                                                                                          },
                                                                                          {
                                                                                              "name":  "remark",
                                                                                              "in":  "query",
                                                                                              "schema":  {
                                                                                                             "type":  "string"
                                                                                                         }
                                                                                          }
                                                                                      ]
                                                                   }
                                                       },
                  "/api/finance/withdraw/{id}/grant":  {
                                                           "put":  {
                                                                       "tags":  [
                                                                                    "Finance"
                                                                                ],
                                                                       "summary":  "发放提现",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "id",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "integer"
                                                                                                         }
                                                                                          }
                                                                                      ]
                                                                   }
                                                       },
...
              }
}